restructuring and syntax fixes
